% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/247_cbf_recruiting.R
\name{get_247_fb_recuriting_ratings}
\alias{get_247_fb_recuriting_ratings}
\title{Get the NCAAFB recuriting ratings for a year from 247sports}
\usage{
get_247_fb_recuriting_ratings(year)
}
\arguments{
\item{year}{}
}
\value{
tibble with columns for team name, year, and rating
}
\description{
Get the NCAAFB recuriting ratings for a year from 247sports
}
