% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_cached_html}
\alias{get_cached_html}
\title{Get cached}
\usage{
get_cached_html(url, query = NULL,
  cache_dir = getOption("EndGame.cache_dir", tempdir()),
  check_cache = T, write_to_cache = T)
}
\arguments{
\item{write_to_cache}{}
}
\value{
contents from the url.
}
\description{
Checks a cache (just a directory) before making a GET
Saves the content if there wasn't a file locally
}
