% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl.R
\name{get_nfl_season}
\alias{get_nfl_season}
\title{Get NFL Seasons}
\usage{
get_nfl_season(season)
}
\arguments{
\item{season}{}
}
\value{
tibble of games
}
\description{
Grab the scores of all games for an NFL season
}
